text_classify_project/algorithm_grpc/abcft_algorithm_grpc/requirements.txt
grpcio==1.4.0
kazoo
pymongo
requests
oss2
qcloud_cos
